Week 15 Code
In this week, I updated the code given to me, changing the background color and making the arrow keys the default. I tried associating a noise with the evil birds, but I did not have enough time. I plan to finish this, as well as building a sort of longer boost than just the jump.